Try to preserve current line when moving from one commit to an adjacent commit…
When moving from one diff to the next diff, then one of the files is
common to the two diffs. So all that we need to do is to:

1. In the old diff, get the common file's line number corresponding to
   the current position.

2. In the new diff, set the position corresponding to the common file's
   line number.

I also need to wait for the vertical scroll bar to be unfrozen before
calling scrollToLine(). Otherwise the new scroll position not be saved
and will be changed by updateScrollBars().